requirements.txt
django
django-cors-headers
django-filter
psycopg2-binary
python-decouple
djangorestframework
djangorestframework-simplejwt
drf-yasg
drf-nested-routers
pillow
gunicorn
celery
redis
